pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "nltk>=3.9.1",
    "numpy>=2.2.3",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "scikit-learn>=1.6.1",
    "streamlit>=1.42.2",
    "twilio>=9.4.6",
]